split cards
This commit removes extracts the "connection by attribute" feature into a separate card. This breaking change was decided to improve the maintainability. It also allows easier adaption in future in case other integrations provide the connections in a different way.
